refactor(callbacks): extract Models tab helpers to _callbacks_models.py
Step 3 of the callbacks.py decomposition (issue #87). Continues the
per-tab module split started by PRs #98 (shared infra) and #99 (US Grid
tab).

## What moves

Three Models / Diagnostics tab helpers extracted from callbacks.py to a
new ``components/_callbacks_models.py``:

* ``_format_metric`` — "N/A"-aware formatter so missing scores never
  display as a misleadingly perfect ``0``.
* ``_models_tab_from_redis`` — Redis fast path building the entire
  Models tab (metrics table + 5 charts) from the scoring-job's cached
  diagnostics payload.
* ``_get_feature_importance`` — pulls real XGBoost importances from the
  model cache when available, otherwise returns a static didactic
  fallback so the chart still renders.

``_empty_figure`` — the cross-tab "no data yet" placeholder — also moves
out of callbacks.py and into ``_callbacks_shared.py`` since it's
genuinely shared across every tab (15+ call sites in callbacks.py
alone). Keeping it in callbacks.py would have created a circular import
between the per-tab modules and the host file.

## Why a separate module

These three helpers are the entire Models-tab data path. Co-locating
them clarifies ownership and lets ``register_callbacks`` shrink down to
the Dash decorator wiring without dragging the helper bodies along for
the ride.

## Public-import surface preserved

``components/callbacks.py`` re-imports each function by name. Every
existing site that does ``from components.callbacks import
_models_tab_from_redis`` (tests + the v1 fallback inside
``register_callbacks``) continues to resolve. No caller-side changes.

## Test patches updated

6 ``@patch("components.callbacks.redis_get")`` decorators inside
``TestModelsTabFromRedis`` rewired to ``components._callbacks_models``
— standard "patch where it's used, not where it's defined." All other
patches on the same target across the file stay as-is because they
exercise functions still living in callbacks.py.

## Size impact

callbacks.py: 6,389 → 6,183 (–206 lines, –3%)
_callbacks_models.py: new, 256 lines
_callbacks_shared.py: 282 → 314 (+32 for _empty_figure)

## Tests

All 1,568 unit tests pass locally. Ruff clean across components/.
